additional happy path test
